refactor: remove the `filter` module
